+++ spider/vjudge/vjudge_test.go
@@ -1,0 +1,66 @@
+package vjudge
+
+import (
+	"XCPCer_board/model"
+	"testing"
+)
+
+//检查函数，若不一致返回1
+func isIntMsgDifferent(funcRet, ans map[string]int) bool {
+	for k, v := range ans {
+		if _, r := funcRet[k]; r == false || v != funcRet[k] {
+			return true
+		}
+	}
+	return false
+}
+func isStringMsgDifferent(funcRet, ans map[string]string) bool {
+	for k, v := range ans {
+		if _, r := funcRet[k]; r == false || v != funcRet[k] {
+			return true
+		}
+	}
+	return false
+}
+
+//判断并输出错误
+func checkIntError(t *testing.T, uid string, tp string, all func(uid string) (map[string]int, error),
+	acInt map[string]int) {
+	if ret, err := all(uid); isIntMsgDifferent(ret, acInt) {
+		if err != nil {
+			t.Errorf("Error of %v in all msg: %v", tp, err)
+		}
+		t.Errorf("Error of %v in all msg\n ret= %v  \nbut the ans is %v", tp, ret, acInt)
+	}
+}
+func checkStrError(t *testing.T, uid string, tp string, all func(uid string) (map[string]string, error),
+	acInt map[string]string) {
+	if ret, err := all(uid); isStringMsgDifferent(ret, acInt) {
+		if err != nil {
+			t.Errorf("Error of %v in all msg: %v", tp, err)
+		}
+		t.Errorf("Error of %v in str msg\n ret= %v  \nbut the ans is %v", tp, ret, acInt)
+	}
+}
+func vjTest(t *testing.T) {
+
+	tp := "vJudge"
+	fc1 := ScrapeAll
+	var uid string
+	var vjInt map[string]int
+
+	uid = model.TestVJIdLYF
+	vjInt = map[string]int{
+		"vj_Person_Last_24_Hours_Pass_Number": 0,
+		"vj_Person_Last_30_Days_Pass_Number":  0,
+		"vj_Person_Last_7_Days_Pass_Number":   0,
+		"vj_Person_Pass_Number":               30,
+	}
+	//开始测试
+	checkIntError(t, uid, tp, fc1, vjInt)
+
+}
+
+func TestVj(t *testing.T) {
+	vjTest(t)
+}
